+++ packages/dts-gen/src/templates/expressions/namespace.ts
@@ -4,14 +4,14 @@ import type { TsExpression } from "./expression";
 export class Namespace implements TsExpression {
   constructor(
     private namespace: string,
-    private typeDefenition: TypeDefinition,
-    private savedTypeDefenition: SavedTypeDefinition
+    private typeDefinition: TypeDefinition,
+    private savedTypeDefinition: SavedTypeDefinition
   ) {}
   tsExpression(): string {
     return `
 declare namespace ${this.namespace} {
-    ${this.typeDefenition.tsExpression()}
-    ${this.savedTypeDefenition.tsExpression()}
+    ${this.typeDefinition.tsExpression()}
+    ${this.savedTypeDefinition.tsExpression()}
 }`.trim();
   }
 }