+++ packages/rest-api-client/rollup.config.mjs
@@ -3,16 +3,7 @@ import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
 import json from "@rollup/plugin-json";
 import { terser } from "rollup-plugin-terser";
-import replace from "rollup-plugin-replace";
 import license from "rollup-plugin-license";
-
-// TODO: After importing JSON module become stable, we can import package.json as follows
-// JSON module: https://github.com/tc39/proposal-json-modules
-// Import Assertions: https://github.com/tc39/proposal-import-assertions
-// import pkgJson from "./package.json" assert { type: "json" };
-import { createRequire } from 'module';
-const pkgJson = createRequire(import.meta.url)("./package.json");
-
 import builtins from "rollup-plugin-node-builtins";
 import globals from "rollup-plugin-node-globals";
 import babel from "@rollup/plugin-babel";
@@ -89,9 +80,6 @@ export default defineConfig({
       ],
     }),
     json(),
-    replace({
-      PACKAGE_VERSION: JSON.stringify(pkgJson.version),
-    }),
     globals(),
     builtins(),
     isProd && terser(),