+++ examples/rest-api-client-demo/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @kintone/rest-api-client bumped from ^4.0.3 to ^4.1.0
+  * devDependencies
+    * @kintone/customize-uploader bumped from ^7.0.3 to ^7.0.4
+
+## [4.0.4](https://github.com/kintone/js-sdk/compare/@kintone/rest-api-client-demo@4.0.3...@kintone/rest-api-client-demo@4.0.4) (2023-07-05)
+
+**Note:** Version bump only for package @kintone/rest-api-client-demo
+
+## [4.0.3](https://github.com/kintone/js-sdk/compare/@kintone/rest-api-client-demo@4.0.2...@kintone/rest-api-client-demo@4.0.3) (2023-06-28)
+
+**Note:** Version bump only for package @kintone/rest-api-client-demo
+
+## [4.0.2](https://github.com/kintone/js-sdk/compare/@kintone/rest-api-client-demo@4.0.1...@kintone/rest-api-client-demo@4.0.2) (2023-06-21)
+
+**Note:** Version bump only for package @kintone/rest-api-client-demo
+
+## [4.0.1](https://github.com/kintone/js-sdk/compare/@kintone/rest-api-client-demo@4.0.0...@kintone/rest-api-client-demo@4.0.1) (2023-06-15)
+
+**Note:** Version bump only for package @kintone/rest-api-client-demo
+
 # [4.0.0](https://github.com/kintone/js-sdk/compare/@kintone/rest-api-client-demo@3.0.41...@kintone/rest-api-client-demo@4.0.0) (2023-06-07)
 
 - build!: drop Node v14 support because of the EOL (#2124) ([ef0e004](https://github.com/kintone/js-sdk/commit/ef0e004b40a518a1b5a3aa5d82446c556c742f02)), closes [#2124](https://github.com/kintone/js-sdk/issues/2124)