+++ packages/dts-gen/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.4](https://github.com/kintone/js-sdk/compare/@kintone/dts-gen@7.0.3...@kintone/dts-gen@7.0.4) (2023-07-05)
+
+### Bug Fixes
+
+- **dts-gen:** remove unnecessary `exclude` from tsconfig ([#2158](https://github.com/kintone/js-sdk/issues/2158)) ([26bd214](https://github.com/kintone/js-sdk/commit/26bd2143ea648f81ba3c2e655485137fc32b10ce))
+
+## [7.0.3](https://github.com/kintone/js-sdk/compare/@kintone/dts-gen@7.0.2...@kintone/dts-gen@7.0.3) (2023-06-28)
+
+**Note:** Version bump only for package @kintone/dts-gen
+
+## [7.0.2](https://github.com/kintone/js-sdk/compare/@kintone/dts-gen@7.0.1...@kintone/dts-gen@7.0.2) (2023-06-21)
+
+**Note:** Version bump only for package @kintone/dts-gen
+
+## [7.0.1](https://github.com/kintone/js-sdk/compare/@kintone/dts-gen@7.0.0...@kintone/dts-gen@7.0.1) (2023-06-15)
+
+### Bug Fixes
+
+- **dts-gen:** validate namespace and typename ([#2137](https://github.com/kintone/js-sdk/issues/2137)) ([cd54866](https://github.com/kintone/js-sdk/commit/cd548666599119b2cc06bab2ca6e3bd319454662))
+
 # [7.0.0](https://github.com/kintone/js-sdk/compare/@kintone/dts-gen@6.1.33...@kintone/dts-gen@7.0.0) (2023-06-07)
 
 - build!: drop Node v14 support because of the EOL (#2124) ([ef0e004](https://github.com/kintone/js-sdk/commit/ef0e004b40a518a1b5a3aa5d82446c556c742f02)), closes [#2124](https://github.com/kintone/js-sdk/issues/2124)